lesson-08

+++ lesson-04-spring-boot/src/main/java/ru/geekbrains/persist/Product.java
@@ -17,12 +17,12 @@ public class Product {
     private String productname;
 
     @Column
-    private BigDecimal price;
+    private Integer price;
 
     public Product() {
     }
 
-    public Product(Long id, String productname, BigDecimal price) {
+    public Product(Long id, String productname, Integer price) {
         this.id = id;
         this.productname = productname;
         this.price = price;
@@ -44,11 +44,11 @@ public class Product {
         this.productname = productname;
     }
 
-    public BigDecimal getPrice() {
+    public Integer getPrice() {
         return price;
     }
 
-    public void setPrice(BigDecimal price) {
+    public void setPrice(Integer price) {
         this.price = price;
     }
 }
